Privatize functions in forward-facing modules. Exemptions made for param checking I guess.
