+++ src/astrohack/_utils/_panel.py
@@ -5,10 +5,73 @@ from numba import njit
 from astrohack._utils._conversion import _convert_unit
 from astrohack._utils._algorithms import _least_squares_fit_block
 from astrohack._utils._constants import plot_types
+from astrohack._utils._io import _load_image_xds
+
+from astrohack._classes.telescope import Telescope
+from astrohack._classes.antenna_surface import AntennaSurface
 
 # global constants
 NPAR = 10
 
+def _panel_chunk(panel_chunk_params):
+    """
+    Process a chunk of the holographies, usually a chunk consists of an antenna over a ddi
+    Args:
+        panel_chunk_params: dictionary of inputs
+    """
+    if panel_chunk_params['origin'] == 'AIPS':
+        inputxds = xr.open_zarr(panel_chunk_params['image_name'])
+        telescope = Telescope(inputxds.attrs['telescope_name'])
+        panel_chunk_params['antenna'] = inputxds.attrs['ant_name']
+
+    else:
+        inputxds = _load_image_xds(panel_chunk_params['image_name'],
+                                   panel_chunk_params['antenna'],
+                                   panel_chunk_params['ddi'],
+                                   dask_load=False)
+
+        inputxds.attrs['AIPS'] = False
+
+        if inputxds.attrs['telescope_name'] == "ALMA":
+            tname = inputxds.attrs['telescope_name']+'_'+inputxds.attrs['ant_name'][0:2]
+            telescope = Telescope(tname)
+        elif inputxds.attrs['telescope_name'] == "EVLA":
+            tname = "VLA"
+            telescope = Telescope(tname)
+        else:
+            raise ValueError('Unsuported telescope {0:s}'.format(inputxds.attrs['telescope_name']))
+
+    surface = AntennaSurface(inputxds, telescope, panel_chunk_params['cutoff'], panel_chunk_params['panel_kind'],
+                             panel_margins=panel_chunk_params['panel_margins'])
+
+    surface.compile_panel_points()
+    surface.fit_surface()
+    surface.correct_surface()
+    
+    xds_name = panel_chunk_params['panel_name'] + '/' + panel_chunk_params['antenna'] + '/' + panel_chunk_params['ddi']
+    xds = surface.export_xds()
+    xds.to_zarr(xds_name, mode='w')
+
+
+def _create_phase_model(npix, parameters, wavelength, telescope, cellxy):
+    """
+    Create a phase model with npix by npix size according to the given parameters
+    Args:
+        npix: Number of pixels in each size of the model
+        parameters: Parameters for the phase model in the units described in _phase_fitting
+        wavelength: Observing wavelength, in meters
+        telescope: Telescope object containing the optics parameters
+        cellxy: Map cell spacing, in meters
+
+    Returns:
+
+    """
+    iNPARameters = _external_to_internal_parameters(parameters, wavelength, telescope, cellxy)
+    dummyphase = np.zeros((npix, npix))
+
+    _, model = _correct_phase(dummyphase, cellxy, iNPARameters, telescope.magnification, telescope.focus,
+                              telescope.surp_slope)
+    return model
 
 def _phase_fitting_block(pols, wavelength, telescope, cellxy, amplitude_image, phase_image, pointing_offset,
                          focus_xy_offsets, focus_z_offset, subreflector_tilt, cassegrain_offset):
