+++ src/astrohack/extract_holog.py
@@ -146,13 +146,12 @@ def extract_holog(
     check_if_file_exists(extract_holog_parms['ms_name'])
     check_if_file_will_be_overwritten(extract_holog_parms['holog_name'],extract_holog_parms['overwrite'])
     check_if_file_will_be_overwritten(extract_holog_parms['point_name'],extract_holog_parms['overwrite'])
-    #######################################
         
     ############# Exstract pointing infromation and save to point.zarr #############
-#    try:
-#        pnt_dict = _load_point_file(extract_holog_parms['point_name'])
-#    except:
-#        pnt_dict = _make_ant_pnt_dict(extract_holog_parms['ms_name'], extract_holog_parms['point_name'], parallel=extract_holog_parms['parallel'])
+    #    try:
+    #        pnt_dict = _load_point_file(extract_holog_parms['point_name'])
+    #    except:
+    #        pnt_dict = _make_ant_pnt_dict(extract_holog_parms['ms_name'], extract_holog_parms['point_name'], parallel=extract_holog_parms['parallel'])
     
     pnt_dict = _extract_pointing(extract_holog_parms['ms_name'], extract_holog_parms['point_name'], parallel=extract_holog_parms['parallel'])
 
@@ -327,7 +326,8 @@ def extract_holog(
     
 
 
-def _check_extract_holog_parms(    ms_name,
+def _check_extract_holog_parms(
+    ms_name,
     holog_obs_dict,
     holog_name,
     point_name,
